Autonomous Possibly Complete
Consolidated the autonomous driving member of chassis so that the
various backward forward left right commands send the values for driving
in mecanum. Added the autonomous commands to the Robot.cpp so that in
autonomous, the robot will go in a square 3 seconds in each direction.
Once autonomous is over, all motion is canceled and teleop is started
up.