Autonomous Commit
Perhaps this shall work. I created a command group so that when
autonomous goes, it will clamp onto a tote or bin, turn 90 degrees
(either left or right depending on where we are situated on the field)
while lifting the tote using pneumatics, and then drive forward into the
scoring zone. Also deleted AutoBackward because we only need to move
forward in this period. Kind of started toying with the speed control
loop for autonomous so that our joystick movements are actually followed
by the chassis.